Deprecate Mac Fluent color setters in favor of newer ColorProviding protocol
